Restore the load path when resetting the local store
When an inconsistent environment is detected (a loaded module changed,
directly or because the way its path is resolved changed), the local
store is reset.

This has the unfortunate side-effect of resetting the Load_path storage.

Before the Load_path module, all caching was done by merlin based on the
value of Config.load_path, outside of the Local_store.

The code has been wrong since the introduction of Load_path: its
internal state was always reset.
However, Merlin frontend was "defensively" repopulating it, which
explained some of the slowdown fixed by PR#1323: most of the time this
had no other effect than scanning the file system twice, but if the
state had been reset since last call, the cache would be repopulated.

The bug was particularly hard to track because there is a background job
in Merlin that remove old entries from the cache: if a file hasn't been
used for 5 minutes, its deps are removed from the cache and its
environment is considered inconsistent, even if nothing has changed.
We would then have to wait more than 5 minutes to reproduce this bug.

This can be tweaked by changing Server.server_accept loop in
ocamlmerlin_server.ml:
-      Mocaml.flush_caches ~older_than:300.0 ();
+      Mocaml.flush_caches ~older_than:15.0 ();
for instance, to set a 15 seconds cache.

Fundamentally, the issue is due to Local_store scope being limited to
the typechecker state, while the load_path was considered as part of the
configuration, at least originally.
That way, resetting the Local_store would set the typer in a fresh state
ready to type with the current configuration.
The introduction of Load_path moved the state out of the global Config
and to the Local_store. Now, resetting the typer also resets the load
path. The code that used to work would now typecheck in an empty
path/environment.