Fix a bug with the support for include functor in signatures
In #7, I observed we need to support the fact that merlin's cache may result in
incremental typing of signatures, in a way analagous to the incremental typing
of structures that can already happen in the real compiler because of the repl.
I copied part of our solution for dealing with that in structures - making
transl_signature take an extra parameter for the signature of what's been
previously typed.  But I also accidentally including that previously typed stuff
in transl_signature's output, which should only be the signature of what's new.
Now I've fixed that, following the same approach we use in the structure case.

Signed-off-by: Chris Casinghino <ccasinghino@janestreet.com>