Call flush_all twice when closing a client connection (#52)
The Merlin server computes responses in response to client requests.  Sometimes, it can
erroneously write a response from a previous client to a later client, causing multiple
responses to be returned.  I've figured out why this happens and fixed it; it's yet
another wacky collision between the Merlin server's somewhat unusual architecture and the
reality of Unix.

Specifically, the output channels might still have data in their buffers when we get to
the next client.  We call flush_all to try and prevent that, but that can fail with an IO
error, leaving the data in the buffer, and causing it to be written out for the next
client.

Ideally we'd just discard the buffered data for any channel that has an IO error, but I
doubt that's possible with the stdlib channels - I haven't looked in detail though.
Instead, we just flush the channels into the server's initial file descriptors.  That's
probably /dev/null or some logfiles, so it's an acceptable place to send that useless
data.

It's somewhat hard to reproduce this behavior, but repeatedly creating a client and then
killing it before sending all its input input is semi-reliable.